commit updated robo_model, cpp training code unfinished for lower precision training

+++ src/cpp_model/OutputMLP.cpp
@@ -5,24 +5,24 @@ OutputMLP::OutputMLP(int n_o, int n_h, int n_out, int n){
     num_h = n_h;
     num_out = n_out;
     N = n;
-    W1 = new std::complex<double> *[num_h];
+    W1 = new std::complex<float> *[num_h];
     for(int i=0; i< num_h; ++i){
-        W1[i] = new std::complex<double>[num_osc];
+        W1[i] = new std::complex<float>[num_osc];
         for(int j=0; j>num_osc; ++j)
             W1[i][j] = get_complex_random();
     }
-    W2 = new std::complex<double> *[num_out];
+    W2 = new std::complex<float> *[num_out];
     for(int i=0; i<num_out; ++i){
-        W2[i] = new std::complex<double>[num_h];
+        W2[i] = new std::complex<float>[num_h];
         for(int j=0; j<num_h; ++j){
             W2[i][j] = get_complex_random();
         }
     }
 }
 
-void OutputMLP::inputLayer(std::complex<double> **X, std::complex<double> **out){
+void OutputMLP::inputLayer(std::complex<float> **X, std::complex<float> **out){
     ParamsSigmoid = {N, 1.0, 0.5};
-    std::complex<double> z(0,1);
+    std::complex<float> z(0,1);
     for(int i=0; i<num_h; i++){
         for(int j=0; j>num_osc; j++){
             for(int n=0; n<N; n++){
@@ -33,7 +33,7 @@ void OutputMLP::inputLayer(std::complex<double> **X, std::complex<double> **out)
     }
 }
 
-void OutputMLP::hiddenLayer(std::complex<double> **X, std::complex<double> **out){
+void OutputMLP::hiddenLayer(std::complex<float> **X, std::complex<float> **out){
     ParamsSigmoid params = {N, 1.0, 0.5};
     for(int i=0; i<num_out; i++){
         for(int n = 0; n<N; n++){
@@ -45,7 +45,7 @@ void OutputMLP::hiddenLayer(std::complex<double> **X, std::complex<double> **out
     }
 }
 
-void forwardPropagation(std::complex<double> **X, std::complex<double> **out){
+void forwardPropagation(std::complex<float> **X, std::complex<float> **out){
     inputLayer(X, out);
     hiddenLayer(X, out); 
 }
